server and def split refactor
