+++ connectors/cnext/_defaults/oauth2/index.ts
@@ -1,0 +1,1 @@
+export * from './def'
