+++ connectors/cnext/googledrive/server.ts
@@ -2,4 +2,4 @@ import {generateConnectorServer} from '../_defaults'
 import {def} from './def'
 
 // TODO: shouldn't have to cast as any, figure out what the right generic should be
-export const server = generateConnectorServer(def as any)
+export const server = generateConnectorServer({def} as any)
