+++ connectors/cnext/index.ts
@@ -1,7 +1,5 @@
-import {
-  def as connectorGoogledrive_def,
-  server as connectorGoogledrive_server,
-} from './googledrive/def'
+import {def as connectorGoogledrive_def} from './googledrive/def'
+import {server as connectorGoogledrive_server} from './googledrive/server'
 
 // generated by _generateConnectorLists.ts. Do not modify by hand
 export type {JsonConnectorDef as ConnectorDef} from './def'
