+++ connectors/cnext/schema.ts
@@ -1,16 +1,17 @@
 import type {ConnectorDef} from '@openint/cdk'
 import {z} from '@openint/util'
-import {JsonConnectorDef, zOauthConnectorConfig} from './def'
+import {zOauthConnectorConfig} from './_defaults/oauth2'
+import {JsonConnectorDef} from './def'
 
 export function generateConnectorDef(def: JsonConnectorDef): ConnectorDef {
   const connectorConfig = () => {
     let schema = zOauthConnectorConfig
-    if (['OAUTH1', 'OAUTH2'].includes(def.auth_type)) {
+    if (['OAUTH1', 'OAUTH2'].includes(def.auth.type)) {
       // QQ: is this the right way to do this?
       // this needs to extend the schema of the json connector
       schema = z.object({
         ...schema.shape,
-        ...(def.connector_config as Record<string, z.ZodTypeAny>),
+        ...(def.auth.connector_config as Record<string, z.ZodTypeAny>),
       })
     }
 
@@ -27,20 +28,20 @@ export function generateConnectorDef(def: JsonConnectorDef): ConnectorDef {
       connectorConfig: connectorConfig(),
       connectionSettings:
         // TODO: consider generalizing this
-        ['OAUTH1', 'OAUTH2', 'OAUTH2CC'].includes(def.auth_type)
+        ['OAUTH1', 'OAUTH2', 'OAUTH2CC'].includes(def.auth.type)
           ? z.object({
               oauth: z.object({
-                credentials: def.connection_settings
-                  ? z.any().parse(def.connection_settings)
+                credentials: def.auth.connection_settings
+                  ? z.any().parse(def.auth.connection_settings)
                   : // note: this needs to extend the schema of the json connector
                     // we need to do this in a way that works for oauth
                     z.object({}),
               }),
               metadata: z.record(z.unknown()).optional(),
             })
-          : def.connection_settings &&
-              Object.keys(def.connection_settings).length > 0
-            ? z.any().parse(def.connection_settings)
+          : def.auth.connection_settings &&
+              Object.keys(def.auth.connection_settings).length > 0
+            ? z.any().parse(def.auth.connection_settings)
             : undefined,
       // TODO: review these 3
       preConnectInput: z.object({
@@ -61,7 +62,7 @@ export function generateConnectorDef(def: JsonConnectorDef): ConnectorDef {
       verticals: def.verticals,
       // TODO: Make this dynamic
       logoUrl: `https://cdn.jsdelivr.net/gh/openintegrations/openint@main/apps/web/public/_assets/logo-google-drive.svg`,
-      authType: def.auth_type,
+      authType: def.auth.type,
     },
   }
 }
